Update readme to fix broken links
Note that the root of the issue is that docs are generated differently from the napari.org homepage structure, hence no proper reference can be established whatsoever and this fix is really a hack to make the link work. A proper fix would be unifying of generation to these docs. But since website rework is undergoing, I propose this fix to have the broken link addressed for now.